// src/app/model/user-recent-transaction-type.enum.ts
export enum UserRecentTransactionType {
    VIEWPROFILE = 'VIEWPROFILE',
    CALLING = 'CALLING',
    LIKEIMAGE = 'LIKEIMAGE',
    SHAREIMAGE = 'SHAREIMAGE',
    SHAREALLIMAGE = 'SHAREALLIMAGE',
    ACCEPTIMAGEBYADMIN = 'ACCEPTIMAGEBYADMIN',
    REJECTIMAGEBYADMIN = 'REJECTIMAGEBYADMIN',
    SENDMESSAGE = 'SENDMESSAGE',
    DATEDELETED = 'DATEDELETED',
    MISSEDCALL= 'MISSEDCALL'
}